move mech_* to hvac package
